Restructure store mod
